requirements.txt
flask
pandas
numpy
matplotlib
seaborn
pillow
scikit-learn
tensorflow
keras
kaggle
